+++ CHANGELOG.md
@@ -23,7 +23,8 @@
 
 ### Added
 
-- tbd
+- TV shows and episodes: If there is more than one network (separated by `,`),
+  they will be stored as separate `<studio>` tags in the NFO file (#1705)
 
 ## 2.10.6 - 2023-12-03